refactor: migrate from Azure to S3 for file storage and update related configurations
