+++ backend/cmd/jobs/env.sample
@@ -36,10 +36,8 @@ REDIS_DATABASE=0
 BIGQUERY_PROJECT=<OPTIONAL>
 BIGQUERY_DATASET=<OPTIONAL>
 
-# Azure file storage (used for images)
-AZURE_STORAGE_ACCOUNT_NAME=
-AZURE_STORAGE_ACCOUNT_KEY=
-AZURE_STORAGE_CONTAINER=images
+# S3 file storage (used for images)
+AWS_DIRECTUS_STORAGE_BUCKET=
 
 # videomanipulator
 VIDEOMANIPULATOR_BASE_URL=http://localhost:8005/
